@@ -38,25 +38,16 @@ public class ViewAndDowloadFIle {
         System.out.println(request.getServletPath());
         String pathImage = request.getServletPath().substring(1);
         byte[] result;
-        try{
-                 result =gcsService.downloadFile(pathImage);
-        }catch(NullPointerException ex){
-            String path = pathImage.substring(pathImage.indexOf("/images"));
-            System.out.println(path);
-            URL url = new URL("https://cloud.mogi.vn"+path);
-            InputStream inputStream = url.openStream();
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            byte[] buffer = new byte[1024];
-            int bytesRead;
-            while ((bytesRead = inputStream.read(buffer)) != -1) {
-                baos.write(buffer, 0, bytesRead);
-            }
-
-            inputStream.close();
-            baos.close();
-
-            result = baos.toByteArray();
-        }
+//        try{
+//                 result =gcsService.downloadFile(pathImage);
+//        }catch(NullPointerException ex){
+        String path = pathImage.substring(pathImage.indexOf("/images"));
+        System.out.println(path);
+        URL url = new URL("https://cloud.mogi.vn" + path);
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        baos.write(url.openStream().readAllBytes());
+        result = baos.toByteArray();
+//    }
         return ResponseEntity.ok().contentType(MediaType.IMAGE_JPEG).body(result);
     }
 }
