@@ -2,8 +2,12 @@ package com.trogiare.model;
 
 import lombok.Data;
 
+import java.time.LocalDateTime;
+
 @Data
 public class Message {
-    private String from;
-    private String text;
+    private String idFrom;
+    private String idRecive;
+    private String message;
+    private LocalDateTime date;
 }
