@@ -3,6 +3,7 @@ package com.trogiare.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
@@ -18,11 +19,12 @@ public class Notification {
         @GenericGenerator(name = "objectid-generator", strategy = "com.trogiare.common.ObjectIDGenerator")
         @Column(unique = true, nullable = false, length = 24)
         private String id;
-        private String receiver;
+        private String receiverId;
         private String message;
-        private String typeNotification;
+        private String content;
+        @CreationTimestamp
         private LocalDateTime timestamp;
-        private boolean isRead;
+        private Boolean isRead = false;
 
 
 }
